Create files for testing, create code class + test
