+++ lib/game.rb
@@ -22,38 +22,38 @@ class Game
   end
   
 
-  def generate_easy_code
-    code_to_break = []
-    colors = ['b', 'g', 'r', 'y']
-      4.times do
-        code_to_break << colors.shuffle
-      end
-    @secret_code = code_to_break.flatten.shuffle[0..3]
-    puts @secret_code.inspect
-    return @secret_code
-  end
+  # def generate_easy_code
+  #   code_to_break = []
+  #   colors = ['b', 'g', 'r', 'y']
+  #     4.times do
+  #       code_to_break << colors.shuffle
+  #     end
+  #   @secret_code = code_to_break.flatten.shuffle[0..3]
+  #   puts @secret_code.inspect
+  #   return @secret_code
+  # end
 
-  def generate_medium_code
-    code_to_break = []
-    colors = ['b', 'g', 'r', 'y', 'o']
-      6.times do
-        code_to_break << colors.shuffle
-      end
-    @secret_code = code_to_break.flatten.shuffle[0..5]
-    puts @secret_code.inspect
-    return @secret_code
-  end
+  # def generate_medium_code
+  #   code_to_break = []
+  #   colors = ['b', 'g', 'r', 'y', 'o']
+  #     6.times do
+  #       code_to_break << colors.shuffle
+  #     end
+  #   @secret_code = code_to_break.flatten.shuffle[0..5]
+  #   puts @secret_code.inspect
+  #   return @secret_code
+  # end
 
-  def generate_hard_code
-    code_to_break = []
-    colors = ['b', 'g', 'r', 'y', 'o', 'w']
-      8.times do
-        code_to_break << colors.shuffle
-      end
-    @secret_code = code_to_break.flatten.shuffle[0..7]
-    puts @secret_code.inspect
-    return @secret_code
-  end
+  # def generate_hard_code
+  #   code_to_break = []
+  #   colors = ['b', 'g', 'r', 'y', 'o', 'w']
+  #     8.times do
+  #       code_to_break << colors.shuffle
+  #     end
+  #   @secret_code = code_to_break.flatten.shuffle[0..7]
+  #   puts @secret_code.inspect
+  #   return @secret_code
+  # end
 
   def get_user_difficulty_choice
     puts "Choose your level of difficulty: (e)asy, (m)edium or (h)ard."
