+++ mastermind_runner.rb
